Switch to vcpkg for dependencies handling (#5)
Now that Yuki-cpp/yuki_docker provides vcpkg enabled images, we're switching dependencies handling to vcpkg, leaving conan behind. This is mostly due to the fact that the integration of vcpkg with cmake seems less intrusive than the conan one.

This commit deletes conan related files, adds a vcpkg manifest and update the CI action to test an run without conan.